Cleanup debugging and cache JDK.

+++ .travis.install_jdk.sh
@@ -14,18 +14,19 @@ if [[ -n "$1" ]]; then
   shift 1
   # Can't source twice from install-jdk.sh in same build so we go ahead and manually set up variables.
   export JAVA_TEST_HOME="$HOME/.jdk/openjdk$FEATURE"
-  bash ./install-jdk.sh -v --feature "$FEATURE" --target "$JAVA_TEST_HOME"
-  # install-jdk.sh downloads to this fixed filename, preventing from downloading again
-  rm -f jdk.tar.gz
+  if [[ ! -d "$JAVA_TEST_HOME" ]]; then
+    bash ./install-jdk.sh --feature "$FEATURE" --target "$JAVA_TEST_HOME"
+    # install-jdk.sh downloads to this fixed filename, preventing from downloading again
+    rm -f jdk.tar.gz
+  fi
 else
   export JAVA_TEST_HOME="$JAVA_HOME"
 fi
 
 NEW_JAVA_HOME="$HOME/.jdk/openjdk11"
-bash ./install-jdk.sh -v --feature 11 --target "$NEW_JAVA_HOME"
-rm -f jdk.tar.gz
+if [[ ! -d "$NEW_JAVA_HOME" ]]; then
+  bash ./install-jdk.sh --feature 11 --target "$NEW_JAVA_HOME"
+  rm -f jdk.tar.gz
+fi
 export JAVA_HOME="$NEW_JAVA_HOME"
 export PATH="$JAVA_HOME/bin:$PATH"
-
-env
-export
